+++ src/test/java/com/aquila/chess/Helper.java
@@ -55,8 +55,8 @@ public class Helper {
     }
 
     static private void checkMCTSTreePoliciesAndValues(final MCTSNode node, final List<String> ret) {
-        float[] policies = node.getCacheValue().getPolicies();
-        float sumPolicies = 0.0F;
+        double[] policies = node.getCacheValue().getPolicies();
+        double sumPolicies = 0.0F;
         for (int i = 0; i < policies.length; i++) sumPolicies += policies[i];
         if (node.getNonNullChildsAsCollection().size() > 0 && sumPolicies < 0.9 || sumPolicies > 1.1) {
             ret.add(String.format("sum of policies should be ~= 1. (sum:%f)", sumPolicies));
